Add LUT transforms. Remove bit-depth from Transforms. Add file bit-depth for selected transforms.

+++ src/OpenColorIO/BitDepthUtils.h
@@ -38,8 +38,10 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 OCIO_NAMESPACE_ENTER
 {
 
-// This is used for normalizing various integer bit-depths relative to float.
-float GetBitDepthMaxValue(BitDepth in);
+// Returns a double since often times a ratio of calls to this function is used
+// to calculate a scale factor and this ratio needs to be done at double
+// precision to avoid slight errors in the scaled values.
+double GetBitDepthMaxValue(BitDepth in);
 
 // True if the bit depth is a float
 bool IsFloatBitDepth(BitDepth in);
