+++ src/OpenColorIO/ops/Lut3D/Lut3DOpCPU.cpp
@@ -351,15 +351,15 @@ BaseLut3DRenderer::~BaseLut3DRenderer()
 
 void BaseLut3DRenderer::updateData(ConstLut3DOpDataRcPtr & lut)
 {
-    m_alphaScale = GetBitDepthMaxValue(lut->getOutputBitDepth())
-                   / GetBitDepthMaxValue(lut->getInputBitDepth());
+    m_alphaScale = (float)(GetBitDepthMaxValue(lut->getOutputBitDepth()))
+                   / (float)GetBitDepthMaxValue(lut->getInputBitDepth());
 
     m_dim = lut->getArray().getLength();
 
     m_maxIdx = (float)(m_dim - 1);
 
     m_step = ((float)m_dim - 1.0f)
-             / GetBitDepthMaxValue(lut->getInputBitDepth());
+             / (float)GetBitDepthMaxValue(lut->getInputBitDepth());
 
 #ifdef USE_SSE
     Platform::AlignedFree(m_optLut);
@@ -1656,9 +1656,9 @@ void InvLut3DRenderer::updateData(ConstLut3DOpDataRcPtr & lut)
     m_tree.initialize(m_grvec.data(), m_dim);
     //m_tree.print();
 
-    float outMax = GetBitDepthMaxValue(lut->getOutputBitDepth());
+    float outMax = (float)GetBitDepthMaxValue(lut->getOutputBitDepth());
 
-    m_alphaScaling = outMax / GetBitDepthMaxValue(lut->getInputBitDepth());
+    m_alphaScaling = outMax / (float)GetBitDepthMaxValue(lut->getInputBitDepth());
 
     // Converts from index units to inDepth units of the original LUT.
     // (Note that inDepth of the original LUT is outDepth of the inverse LUT.)
@@ -1668,7 +1668,7 @@ void InvLut3DRenderer::updateData(ConstLut3DOpDataRcPtr & lut)
 
     // TODO: Should improve this based on actual LUT contents since it
     // is legal for LUT contents to exceed the typical scaling range.
-    m_inMax = GetBitDepthMaxValue(lut->getInputBitDepth());
+    m_inMax = (float)GetBitDepthMaxValue(lut->getInputBitDepth());
 }
 
 void InvLut3DRenderer::extrapolate3DArray(ConstLut3DOpDataRcPtr & lut)
@@ -1698,7 +1698,7 @@ void InvLut3DRenderer::extrapolate3DArray(ConstLut3DOpDataRcPtr & lut)
         }
     }
 
-    const float center = GetBitDepthMaxValue(depth) * 0.5f;
+    const float center = (float)GetBitDepthMaxValue(depth) * 0.5f;
     const float scale = 4.f;
 
     // Extrapolate faces.
